pyproject.toml
[project]
name = "langgraph-task-orchestrator"
version = "0.1.0"
description = "Multi-agent OKR analytics system with LangGraph, HITL, and LLM-as-Judge evaluation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-groq>=0.1.0",
    "langchain-community>=0.2.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "chromadb>=0.5.0",
    "qdrant-client>=1.9.0",
    "ragas>=0.1.0",
    "deepeval>=0.21.0",
    "langfuse>=2.0.0",
    "mlflow>=2.13.0",
    "supabase>=2.4.0",
    "streamlit>=1.35.0",
]
[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88